editnoterendererphone, maingamerendererphonerpl: fix alpha types and values
Disassembly audit of the phone renderers (issues Tatsh/me#338 and #339).

- EditNoteRendererPhone renderMusicBar:timeline:alpha: (0x21079c) and
  renderTuneInfo:artworkSize:alpha: (0x210b48) declared alpha as float, but the
  method type encodings are 'd' (v44@0:8{CGPoint=dd}16B32d36 and
  v48@0:8{CGPoint=dd}16d32d40). Widen alpha to double in the header and impl,
  matching the encoding and the EditNoteRendererPad/MainGameRendererPad siblings.
- MainGameRendererPhoneRpl upper-beam block (0x14c5d4): both
  drawInRect:fromRegion:transform:alpha: calls passed alpha 0.0f, which would draw
  nothing. The binary loads 0x3f000000 (0.5f) into w26 and stores it to the
  stack-passed alpha slot before both calls (0x14c8ec/0x14c904/0x14c9d0); restore 0.5f.

Both verified against the arm64 disassembly.

+++ Project/EditNoteRendererPhone.h
@@ -137,7 +137,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @param alpha The opacity.
  * @ghidraAddress 0x21079c
  */
-- (void)renderMusicBar:(CGPoint)position timeline:(BOOL)timeline alpha:(float)alpha;
+- (void)renderMusicBar:(CGPoint)position timeline:(BOOL)timeline alpha:(double)alpha;
 
 /**
  * @brief Draws the tune information block: jacket, title, difficulty word, and level.
@@ -146,7 +146,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @param alpha The opacity.
  * @ghidraAddress 0x210b48
  */
-- (void)renderTuneInfo:(CGPoint)position artworkSize:(double)artworkSize alpha:(float)alpha;
+- (void)renderTuneInfo:(CGPoint)position artworkSize:(double)artworkSize alpha:(double)alpha;
 
 /**
  * @brief Draws the upper background. This phone override does nothing.
